move "function_has_named_arguments" + add test with a functor
